carbon2185/requirements.txt
psycopg2-binary
colorama
